move chain related info to new file
