+++ src/index.ts
@@ -1,3 +1,5 @@
+export * from './addresses'
+export * from './chains'
 export * from './constants'
 export * from './entities'
 export * from './utils'
